move relations to traits
